Move the AST test class into it's own file. Also fixed the get_numpr_prop method
